Switch mob health from int to byte

+++ src/game/mob/Human.java
@@ -131,7 +131,7 @@ public class Human {
     };
 
     public static void registerHumanMob(){
-        registerMob(new MobDefinition("human", "hurt", 7, bodyMeshes, bodyOffsets, bodyRotations,1.9f, 0.25f, mobInterface));
+        registerMob(new MobDefinition("human", "hurt", (byte) 7, bodyMeshes, bodyOffsets, bodyRotations,1.9f, 0.25f, mobInterface));
     }
